+++ site/pub/examples/level.js
@@ -14,6 +14,7 @@ loadSound("coin", "coin.ogg");
 init({
 	fullscreen: true,
 	scale: 2,
+	debug: true,
 });
 
 scene("main", () => {
@@ -150,16 +151,10 @@ scene("main", () => {
 	});
 
 	const score = add([
-		text("0"),
 		pos(12, 12),
 		layer("ui"),
 		{ value: 0, },
 	]);
-
-	keyPress("f1", () => {
-		kaboom.debug.showArea = !kaboom.debug.showArea;
-		kaboom.debug.hoverInfo = !kaboom.debug.hoverInfo;
-	});
 
 });